+++ exceptions/RuntimeException.d.ts
@@ -1,4 +1,4 @@
 export default class RuntimeException extends Error {
     code: number;
-    constructor(message?: string, code?: number);
+    constructor(message?: string, code?: number | undefined | null, stack?: string);
 }