+++ server.js
@@ -10,24 +10,24 @@ let ExceptionHandler;
 try {
     ExceptionHandler = require(exceptionHandlerPath).default;
 }
-catch {
-    throw new RuntimeException(`Missing exception handler class [${exceptionHandlerPath}].`);
+catch (error) {
+    throw new RuntimeException(`Missing exception handler class [${exceptionHandlerPath}].`, null, error.message);
 }
 const apiRoutesPath = App.Path.routesPath("api.ts");
 let ApiRoutes;
 try {
     ApiRoutes = require(apiRoutesPath).default;
 }
-catch {
-    throw new RuntimeException(`Missing api file on routes directory [${apiRoutesPath}].`);
+catch (error) {
+    throw new RuntimeException(`Missing api file on routes directory [${apiRoutesPath}].`, null, error.message);
 }
 const webRoutesPath = App.Path.routesPath("web.ts");
 let WebRoutes;
 try {
     WebRoutes = require(webRoutesPath).default;
 }
-catch {
-    throw new RuntimeException(`Missing web file on routes directory [${webRoutesPath}].`);
+catch (error) {
+    throw new RuntimeException(`Missing web file on routes directory [${webRoutesPath}].`, null, error.message);
 }
 const server = Bun.serve({
     development: Bun.env.NODE_ENV !== "production" && {