+++ src/builders/StorageBuilder.ts
@@ -1,6 +1,7 @@
 import App from "@bejibun/app";
 import {isEmpty} from "@bejibun/utils";
 import Luxon from "@bejibun/utils/facades/Luxon";
+import Str from "@bejibun/utils/facades/Str";
 import path from "path";
 
 export default class StorageBuilder {
@@ -10,6 +11,7 @@ export default class StorageBuilder {
 
     public constructor() {
         this.name = "";
+        this.directory = "general";
     }
 
     public setFile(file: any): StorageBuilder {
@@ -31,8 +33,8 @@ export default class StorageBuilder {
     }
 
     public async save(): Promise<any> {
-        if (isEmpty(this.name)) this.name = `${Luxon.DateTime.now().toUnixInteger()}-${Math.random().toString(36).substring(2, 2 + 16)}`;
+        if (isEmpty(this.name)) this.name = Str.random();
 
-        await Bun.write(App.Path.storagePath(`app/public/${this.directory}/${this.name}${path.extname(this.file?.name)}`), this.file);
+        await Bun.write(App.Path.storagePath(`app/public/${this.directory}/${Luxon.DateTime.now().toUnixInteger()}-${this.name}${path.extname(this.file?.name)}`), this.file);
     }
 }